refactor(async-react): consolidate MSW debug contract for wave 2
- Collapse EndpointLatency to {mode, ms} and add parseEndpointLatency
- Move debugging contract to src/mocks/debugging.ts
- Trim debug POST response; seed debugger from getDebuggingState()
- Drop response:mocked listener; type toggle params; dedupe progress CSS
- README: explain debug channel as HTTP for visible config requests
